Extract paragraph editor
