+++ src/editors/passage/index.less
@@ -20,6 +20,12 @@
 	}
 }
 
+.passage-edior {
+	display: flex;
+	flex-direction: column;
+	height: 100%;
+}
+
 ul.CodeMirror-hints {
 	position: absolute;
 	list-style-type: none;
