+++ src/story-edit-view/passage-list-view/index.html
@@ -1,0 +1,3 @@
+<modal-dialog id="passagesListModal" :origin="origin" :can-widen="true" @destroyed="dialogDestroyed()">
+
+</modal-dialog>
